+++ routes/parametros/parametros.contacto.route.js
@@ -5,11 +5,13 @@ const ListaDePrecios = require("../../models/listaDePrecios.model")
 
 
 app.post("/", async (req, res, next) => {
-  ["domicilios", "contactos", "cuentas", "eliminado", "rutas"].forEach(x => {
+  ["contactos", "cuentas", "eliminado", "rutas"].forEach(x => {
     if (Object.keys(req.body).includes(x)) delete req.body[x]
   })
-
+  
   // Obtenemos las listas existentes para obtener el id.
+
+   req.body = organizarDomicilios(req.body)
 
   let listasExistentes = await ListaDePrecios.find().select("nombre").exec()
   
@@ -24,6 +26,7 @@ app.post("/", async (req, res, next) => {
       )
       .find()
       .select('_id nombre email')
+      .lean()
       .exec()
     // Transformamos las etiquetas.
 
@@ -36,8 +39,7 @@ app.post("/", async (req, res, next) => {
         .filter(x => x)
     }
 
-    // Remplazar listas
-
+    // Remplazar lista
     if (datos?.listaDePrecios) {
       let idLista = listasExistentes.find(
         x => x.nombre === datos.listaDePrecios
@@ -46,15 +48,14 @@ app.post("/", async (req, res, next) => {
       if (!idLista) {
         error = `No existe una lista llamada ${datos.listaDePrecios}`
       } else datos.listaDePrecios = idLista
-    } else datos["listaDePrecios"] = undefined
-
+    }
 
     // Agregar vendedores
     // Debe ser un string separado por comas
     if (datos?.usuariosAsignados) {
-      
+
       let leyenda = ' [ NO ENCONTRADO ]'
-      let msjError = (d) => `${leyenda} "${d}"`
+      let msjError = (d) => `${ leyenda } "${ d }"`
       
       let usuariosConvertidos = datos.usuariosAsignados
         .split(',')
@@ -70,8 +71,6 @@ app.post("/", async (req, res, next) => {
         
     } else datos['usuariosAsignados'] = []
  
-
-    
     let filter = {
       codigo: datos.codigo,
     }
@@ -108,13 +107,76 @@ app.post("/", async (req, res, next) => {
         respuesta.filter(x => x.status === "rejected").map(x => x.reason) ?? []
       let correctos =
         respuesta.filter(x => x.status === "fulfilled")?.length ?? 0
-
-      res.send({
-        rechazados,
-        correctos,
+      
+       req.version_offline().then(parametros => {
+         res.send({
+           rechazados,
+           correctos,
+           version:parametros.version_offline 
+        })
       })
+      .catch(_=>next(_))
     })
     .catch(_ => next(_))
 })
+
+
+
+
+/**
+ *Devuevle los domicilios estructurados
+ *Debe tener la estructura 
+ *  domicilio1_numeroExterior
+ *  domicilio1_numeroInterior
+ *  domicilio1_calle
+ * 
+ *  domicilio2_calle
+ *
+ * 
+ * Los numeros depues de domicilio son los que agrupan el domicilio
+ * @param {*} datosEstructurados
+ */
+function organizarDomicilios(datosEstructurados) {
+  
+  let nombreEncabezado = "domicilio"
+  
+  if (datosEstructurados?.length < 1) return []
+  //Debe haber por lo menos un encabezado con "domicilio" 
+  let todosLosEncabezados = Object.keys(datosEstructurados[0])
+  let existeDomicilio = !!todosLosEncabezados
+  .find(x => x.includes(nombreEncabezado))
+  if (!existeDomicilio) return []
+
+  // Obtenemos todos los encabezados que incluyen domicilio
+  let encabezados = todosLosEncabezados.filter(x => x.includes(nombreEncabezado))
+  let re = /\D+/g
+  let cantidadDeDomicilios = Array.from(
+    
+    new Set(encabezados.map(x => x.replace(re, "")))
+  
+  )
+
+  return datosEstructurados
+    .map(linea => {
+      linea['domicilios'] = []
+      cantidadDeDomicilios.forEach(i => {
+        let domicilio = {}
+       //recorremos todos los encabezados de domicilios existenes 
+        encabezados.forEach(encabezado => { 
+
+        //Debe de incluir el grupo actual i
+          let grupoActual = nombreEncabezado + i + "_"
+          if (encabezado.includes(grupoActual)) {
+            
+            let variable = encabezado.slice(grupoActual.length)
+            domicilio[variable] = linea[encabezado]
+          }
+        })
+
+        linea['domicilios'].push(domicilio)
+  })
+  return linea
+  })
+}
 
 module.exports = app